+++ Runtime/AssemblyInfo.cs
@@ -1,6 +1,6 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo(
-    "Deucarian.ViewerAuthentication.Editor")]
+    "Deucarian.Authentication.Editor")]
 [assembly: InternalsVisibleTo(
-    "Deucarian.ViewerAuthentication.Tests.EditMode")]
+    "Deucarian.Authentication.Tests.EditMode")]